contracts-bedrock: fix Encoding.sol

+++ packages/contracts-bedrock/src/libraries/Encoding.sol
@@ -215,7 +215,7 @@ library Encoding {
             blobBaseFee,
             hash,
             batcherHash,
-            dependencySet.length,
+            uint8(dependencySet.length),
             dependencySet
         );
     }
